Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
bcrypt==4.1.2
cryptography==42.0.2
pytz==2024.1
dnspython==2.6.1
pymongo==4.6.2
motor==3.3.2
python-dotenv==1.0.1